Rework MUI V5 test suite to use CRA Use create-react-app instead of storybook for better performance
[ghstack-poisoned]

+++ scripts/postBuild.js
@@ -1,8 +1,14 @@
 const fs = require('fs');
 const path = require('path');
 
+const skipPostBuildPackages = ['e2e-mui-v5'];
+
 function postBuild(dir) {
   // Rename package/build/src to package/dist
+  if (skipPostBuildPackages.some((pkg) => dir.includes(pkg))) {
+    return;
+  }
+
   const dist = path.join(dir, 'dist');
   if (fs.existsSync(dist)) {
     fs.rmSync(dist, { recursive: true });
